split inspector strategies into multiple files
