+++ editor/src/components/inspector/inspector-strategies/inspector-strategies.tsx
@@ -1,24 +1,20 @@
-import { ElementInstanceMetadataMap } from '../../core/shared/element-template'
-import { ElementPath } from '../../core/shared/project-file-types'
-import * as PP from '../../core/shared/property-path'
-import { CanvasCommand } from '../canvas/commands/commands'
-import { setProperty } from '../canvas/commands/set-property-command'
-import { EditorDispatch } from '../editor/action-types'
+import { ElementInstanceMetadataMap } from '../../../core/shared/element-template'
+import { ElementPath } from '../../../core/shared/project-file-types'
+import * as PP from '../../../core/shared/property-path'
+import { CanvasCommand } from '../../canvas/commands/commands'
+import { setProperty } from '../../canvas/commands/set-property-command'
+import { EditorDispatch } from '../../editor/action-types'
 import {
   fillContainerApplicable,
   filterKeepFlexContainers,
   FlexAlignment,
   FlexJustifyContent,
   hugContentsApplicable,
-} from './inspector-common'
-import { applyCommandsAction } from '../editor/actions/action-creators'
-import { deleteProperties } from '../canvas/commands/delete-properties-command'
-import { CSSNumber, FlexDirection, printCSSNumber } from './common/css-utils'
-
-export type InspectorStrategy = (
-  metadata: ElementInstanceMetadataMap,
-  selectedElementPaths: Array<ElementPath>,
-) => Array<CanvasCommand> | null
+} from '../inspector-common'
+import { applyCommandsAction } from '../../editor/actions/action-creators'
+import { deleteProperties } from '../../canvas/commands/delete-properties-command'
+import { CSSNumber, FlexDirection, printCSSNumber } from '../common/css-utils'
+import { InspectorStrategy } from './inspector-strategy'
 
 export const setFlexAlignJustifyContentStrategies = (
   flexAlignment: FlexAlignment,
@@ -130,18 +126,3 @@ export const setPropHugStrategies = (prop: 'width' | 'height'): Array<InspectorS
     )
   },
 ]
-
-export function runStrategies(
-  dispatch: EditorDispatch,
-  metadata: ElementInstanceMetadataMap,
-  selectedViews: ElementPath[],
-  strategies: InspectorStrategy[],
-): void {
-  for (const strategy of strategies) {
-    const commands = strategy(metadata, selectedViews)
-    if (commands != null) {
-      dispatch([applyCommandsAction(commands)])
-    }
-    return
-  }
-}
