+++ editor/src/components/inspector/inspector-strategies/inspector-strategy.tsx
@@ -1,0 +1,25 @@
+import { ElementInstanceMetadataMap } from '../../../core/shared/element-template'
+import { ElementPath } from '../../../core/shared/project-file-types'
+import { CanvasCommand } from '../../canvas/commands/commands'
+import { EditorDispatch } from '../../editor/action-types'
+import { applyCommandsAction } from '../../editor/actions/action-creators'
+
+export type InspectorStrategy = (
+  metadata: ElementInstanceMetadataMap,
+  selectedElementPaths: Array<ElementPath>,
+) => Array<CanvasCommand> | null
+
+export function runStrategies(
+  dispatch: EditorDispatch,
+  metadata: ElementInstanceMetadataMap,
+  selectedViews: ElementPath[],
+  strategies: InspectorStrategy[],
+): void {
+  for (const strategy of strategies) {
+    const commands = strategy(metadata, selectedViews)
+    if (commands != null) {
+      dispatch([applyCommandsAction(commands)])
+    }
+    return
+  }
+}
